[Feat]#11- GroupListViewCell의 확장을 파일로 분리
- extension을 파일을 만들어 작성하였습니다.
